requirements.txt
youtube_transcript_api
streamlit
google-generativeai
python-dotenv
pathlib
langchain-google-genai
langchain
langchain_community